Update Database Functionality and Bug Fix
Added database functionality to Services and Maps. Only missing the search bar. Closes #22 and closes #23.

Updated NSUserDefaults to CampusId. Closes #19

A typo caused multiple errors with optionals and led me to believe int32 type variables were the root of the problem. They. Were. Not. Closes #25

Renamed columns in order to avoid more errors. Closes #26

Updated the database string with an extra '/' in order to avoid possible errors with the file path. Closes #21.

Closing Nichola's Test Issue. Closes #15.

That should be all bug fixing for this commit.