Check Web Service Connection
Added Web Service Connection check to the WebServicesHelper file. Closes #56.

Added required note to the property list file in order to transit data to and from the application.

Added check for lack of rooms on a specific campus for the database (avoid crashing the application if there are no rooms). Closes #51.