Update Web Services and Queues
Updated multiple classes and web services.

Added the version field to the campus table and to the model as well. Closes #66.

Removed code made specifically for iOS7 and lower versions as the target for the project has been set to iOS 8.0. Closes #68.

Updated database manager in order to add buildings and rooms properly. Commented out test data for the application, all data in there is now from the server.

Updated database statements in order to conform with the Campus Version fields.

Updated Splash, Select Campus, Main Menu and Services controllers in order to use the dispatch methods and threads. Forced the main thread to sleep in order to test the application loading time with the web services. Need to update this to display the page, show a spinner and realize the requests. Closes #65.

Updated MapsViewController in order to not make an extra, unnecessary call to the database. Changed a few variables around in order to reduce the amount of data being past from a controller to the other.

Updated a few methos in WebServicesHelper in order to correctly get rooms and buildings. Also updated the XMLParsers class for the respective methods. Still need to complete buildings when the web service method for the building is updated. Closes #59 and closes #60.