Update General Application
Updated the building parser in order to retrieve the essential information out of the SOAP object. Closes #72.

Fixed a bug where the model objects would be null due to the fact that the web service did not complete on time. Closes #71.

Fixed the positions for labels on the PostMapsViewController.

Added a few icons to the projects, started linking them to the actual design.