General Update - Pre-release 0.9
Fixed image issues in the web service request. Closes #90.

Added support for multiple images in the indoor maps section as well as splitting the title in order to label the segments properly. Closes #99.

Fixed spacing issue in campus selection related to the programatically generated navigation bar. Closes #95.

Fixed activity indicator weird positioning throughout the app to be on the center of the tableView/ViewController. Closes #96.

Added check for campuses on the splash screen in order to avoid redundant calls to the database. Need to implement a web service method to check if new campus were added. Closes #97.

Updated web service helper in order to properly retrieve and store images for the current search as well as removed all the redundant use of the buildings in the database. Table still exists if it is required in the future. Closes #98, closes #101.

Fixed Services/MainMenu eternal wait when a building failed to download. Closes #100.

Fixed the positioning and the information for the labels on the PostMaps page. Closes #103.

Added missing orange-colored disclosure indicators to the application. Closes #104, this also completes the current design placed inside the application code. Closes #28.

Fixed swipe which would allow user to return to the splash screen and recall the campuses download and other methods. Closes #105.

Fixed tab disappearance during the first use by moving segues around and removing a redundant navigation controller. Closes #106.

Fixed half of the campus selection back button error. Persisting only during first use. Closes #107.

Removed test data and most print statements from the application, searching for the missing ones. Closes #102.

Pre-release 0.9 for Central Wayfinder.