Update Indoor Maps and Campus Selection
Application now loads the indoor maps in reverse, showing floors from 1 to the last available.

Fixed a few issues with loading the indoor maps and deleting images from the web service, such as the images being deleted prior to completion of the downloads.

Fixed Campus Selection Thread issues and added proper handling for the multiple uses of the page.